+++ clientgui/ValidateAccountKey.cpp
@@ -112,8 +112,8 @@ bool CValidateAccountKey::TransferFromWindow(void) {
 
 
 bool CValidateAccountKey::wxIsAlphaNumeric(const wxString& val) {
-    int i;
-    for (i = 0; i < (int)val.Length(); i++) {
+    size_t i;
+    for (i = 0; i < val.Length(); i++) {
         if (!wxIsalnum(val[i]))
             return FALSE;
     }
